Fix MSVC build problems.

+++ interaction/DragWatcher.cpp
@@ -42,7 +42,7 @@ DragWatcher::haveSignificantDrag() const
 		msec_passed += 60*60*24;
 	}
 
-	double const dist_score = sqrt(m_dragMaxSqDist) / 12.0;
+	double const dist_score = sqrt((double)m_dragMaxSqDist) / 12.0;
 	double const time_score = msec_passed / 500.0;
 
 	return dist_score + time_score >= 1.0;
